= SipUserAgent =  * doxygen 을 위한 주석을 추가함.
= TestSipParser =
 * 64bit 에서 warning 없이 컴파일되도록 수정함.

+++ trunk/SipClient/SipClientMain.cpp
@@ -1,3 +1,21 @@
+/* 
+ * Copyright (C) 2012 Yee Young Han <websearch@naver.com> (http://blog.naver.com/websearch)
+ *
+ * This program is free software; you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation; either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program; if not, write to the Free Software
+ * Foundation, Inc., 59 Temple Place, Suite 330, Boston, MA  02111-1307  USA 
+ */
+
 #include "SipClient.h"
 
 extern std::string	gstrInviteId;
@@ -59,7 +77,7 @@ int main( int argc, char * argv[] )
 	{
 		if( szCommand[0] == 'Q' || szCommand[0] == 'q' ) break;
 
-		iLen = strlen(szCommand);
+		iLen = (int)strlen(szCommand);
 		if( iLen >= 2 && szCommand[iLen-2] == '\r' )
 		{
 			szCommand[iLen-2] = '\0';
